CodeGuardian_AI/requirements.txt
pyautogen
python-dotenv
google-generativeai